requirements.txt
absl-py==1.0.0
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attr==0.3.2
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==21.12b0
bleach==4.1.0
blinker==1.7.0
build==1.0.3
cachetools==5.0.0
certifi==2024.2.2
cffi==1.15.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.23
click==8.0.3
coloredlogs==15.0.1
coremltools==4.0b3
cycler==0.11.0
dataclasses-json==0.6.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diskcache==5.6.3
exceptiongroup==1.2.0
executing==0.8.2
fastapi==0.109.2
filelock==3.13.1
flatbuffers==2.0
fonttools==4.29.1
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.62.0
graphlib-backport==1.0.3
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
h5py==3.6.0
httptools==0.6.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kubernetes==29.0.0
langchain==0.1.9
langchain-community==0.0.22
langchain-core==0.1.26
langsmith==0.1.5
libclang==13.0.0
llama_cpp_python==0.2.50
llvmlite==0.33.0
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
mmh3==4.1.0
monotonic==1.6
mpmath==1.2.1
multidict==6.0.5
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==3.1
notebook==6.4.8
numba==0.50.1
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.17.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opt-einsum==3.3.0
orjson==3.9.14
overrides==7.7.0
packaging==23.2
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
pkgutil_resolve_name==1.3.10
platformdirs==2.4.1
posthog==3.4.2
potracer==0.0.4
prettytable==0.7.2
prometheus-client==0.13.0
prompt-toolkit==3.0.24
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.0.7
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2021.3
PyYAML==6.0.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.0
resampy==0.2.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.8
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.0
sentence-transformers==2.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.1.4
starlette==0.36.3
streamlit==1.31.1
sympy==1.9
tenacity==8.2.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.13.0
testpath==0.5.0
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
tomli==1.2.3
toolz==0.12.1
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tornado==6.1
tqdm==4.66.2
traitlets==5.1.1
transformers==4.38.1
triton==2.2.0
turicreate==6.4.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
validators==0.22.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.9.4
zipp==3.17.0
